- Created a Node Manager
- Node Manager Prefab Created
- Node Manager object will keep track of any nodes you give it in the
inspector
- If all of these nodes all lit (isLit), then the nodemanager can send a
message to other scripts to do whatever you'd like it to do. For
instance, if you want a door to open when the conditions are met, but
close when the torches are removed, just type in the methods you'd like
to call in the NodeManger inspector window!
- Added a Door control script to control a simple door opening/closing
animation or position change